docs: Update examples
