+++ packages/hub-nodejs/examples/README.md
@@ -3,7 +3,7 @@
 A collection of Typescript examples showcasing the things you can do with `hub-nodejs`. 
 
 - [Generate a chronological feed](./chron-feed/)
-- [Writing data for a user](./write-data/)
-- [Building casts correctly](./make-cast/)
+- [Creating an account](hello-world/)
+- [Writing different types of messages to the hub](write-data/)
 
 Contributions for other examples are welcome!
