+++ packages/hub-nodejs/examples/hello-world/tsconfig.json
