--- packages/hub-nodejs/examples/write-data/abi/StorageRegistryABI.ts
@@ -1,579 +1,0 @@
-export const StorageRegistryABI = [
-  {
-    inputs: [
-      { internalType: "contract AggregatorV3Interface", name: "_priceFeed", type: "address" },
-      { internalType: "contract AggregatorV3Interface", name: "_uptimeFeed", type: "address" },
-      { internalType: "uint256", name: "_initialUsdUnitPrice", type: "uint256" },
-      { internalType: "uint256", name: "_initialMaxUnits", type: "uint256" },
-      { internalType: "address", name: "_initialVault", type: "address" },
-      { internalType: "address", name: "_initialRoleAdmin", type: "address" },
-      { internalType: "address", name: "_initialOwner", type: "address" },
-      { internalType: "address", name: "_initialOperator", type: "address" },
-      { internalType: "address", name: "_initialTreasurer", type: "address" },
-    ],
-    stateMutability: "nonpayable",
-    type: "constructor",
-  },
-  { inputs: [], name: "CallFailed", type: "error" },
-  { inputs: [], name: "ContractDeprecated", type: "error" },
-  { inputs: [], name: "ExceedsCapacity", type: "error" },
-  { inputs: [], name: "GracePeriodNotOver", type: "error" },
-  { inputs: [], name: "IncompleteRound", type: "error" },
-  { inputs: [], name: "InvalidAddress", type: "error" },
-  { inputs: [], name: "InvalidAmount", type: "error" },
-  { inputs: [], name: "InvalidBatchInput", type: "error" },
-  { inputs: [], name: "InvalidDeprecationTimestamp", type: "error" },
-  { inputs: [], name: "InvalidFixedPrice", type: "error" },
-  { inputs: [], name: "InvalidMaxAnswer", type: "error" },
-  { inputs: [], name: "InvalidMinAnswer", type: "error" },
-  { inputs: [], name: "InvalidPayment", type: "error" },
-  { inputs: [], name: "InvalidPrice", type: "error" },
-  { inputs: [], name: "InvalidRangeInput", type: "error" },
-  { inputs: [], name: "InvalidRoundTimestamp", type: "error" },
-  { inputs: [], name: "NotOperator", type: "error" },
-  { inputs: [], name: "NotOwner", type: "error" },
-  { inputs: [], name: "NotTreasurer", type: "error" },
-  { inputs: [], name: "PriceOutOfBounds", type: "error" },
-  { inputs: [], name: "SequencerDown", type: "error" },
-  { inputs: [], name: "StaleAnswer", type: "error" },
-  { inputs: [], name: "Unauthorized", type: "error" },
-  {
-    anonymous: false,
-    inputs: [{ indexed: false, internalType: "address", name: "account", type: "address" }],
-    name: "Paused",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: true, internalType: "address", name: "payer", type: "address" },
-      { indexed: true, internalType: "uint256", name: "fid", type: "uint256" },
-      { indexed: false, internalType: "uint256", name: "units", type: "uint256" },
-    ],
-    name: "Rent",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: true, internalType: "bytes32", name: "role", type: "bytes32" },
-      { indexed: true, internalType: "bytes32", name: "previousAdminRole", type: "bytes32" },
-      { indexed: true, internalType: "bytes32", name: "newAdminRole", type: "bytes32" },
-    ],
-    name: "RoleAdminChanged",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: true, internalType: "bytes32", name: "role", type: "bytes32" },
-      { indexed: true, internalType: "address", name: "account", type: "address" },
-      { indexed: true, internalType: "address", name: "sender", type: "address" },
-    ],
-    name: "RoleGranted",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: true, internalType: "bytes32", name: "role", type: "bytes32" },
-      { indexed: true, internalType: "address", name: "account", type: "address" },
-      { indexed: true, internalType: "address", name: "sender", type: "address" },
-    ],
-    name: "RoleRevoked",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: false, internalType: "uint256", name: "oldDuration", type: "uint256" },
-      { indexed: false, internalType: "uint256", name: "newDuration", type: "uint256" },
-    ],
-    name: "SetCacheDuration",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: false, internalType: "uint256", name: "oldTimestamp", type: "uint256" },
-      { indexed: false, internalType: "uint256", name: "newTimestamp", type: "uint256" },
-    ],
-    name: "SetDeprecationTimestamp",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: false, internalType: "uint256", name: "oldPrice", type: "uint256" },
-      { indexed: false, internalType: "uint256", name: "newPrice", type: "uint256" },
-    ],
-    name: "SetFixedEthUsdPrice",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: false, internalType: "uint256", name: "oldPeriod", type: "uint256" },
-      { indexed: false, internalType: "uint256", name: "newPeriod", type: "uint256" },
-    ],
-    name: "SetGracePeriod",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: false, internalType: "uint256", name: "oldAge", type: "uint256" },
-      { indexed: false, internalType: "uint256", name: "newAge", type: "uint256" },
-    ],
-    name: "SetMaxAge",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: false, internalType: "uint256", name: "oldPrice", type: "uint256" },
-      { indexed: false, internalType: "uint256", name: "newPrice", type: "uint256" },
-    ],
-    name: "SetMaxAnswer",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: false, internalType: "uint256", name: "oldMax", type: "uint256" },
-      { indexed: false, internalType: "uint256", name: "newMax", type: "uint256" },
-    ],
-    name: "SetMaxUnits",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: false, internalType: "uint256", name: "oldPrice", type: "uint256" },
-      { indexed: false, internalType: "uint256", name: "newPrice", type: "uint256" },
-    ],
-    name: "SetMinAnswer",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: false, internalType: "uint256", name: "oldPrice", type: "uint256" },
-      { indexed: false, internalType: "uint256", name: "newPrice", type: "uint256" },
-    ],
-    name: "SetPrice",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: false, internalType: "address", name: "oldFeed", type: "address" },
-      { indexed: false, internalType: "address", name: "newFeed", type: "address" },
-    ],
-    name: "SetPriceFeed",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: false, internalType: "address", name: "oldFeed", type: "address" },
-      { indexed: false, internalType: "address", name: "newFeed", type: "address" },
-    ],
-    name: "SetUptimeFeed",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: false, internalType: "address", name: "oldVault", type: "address" },
-      { indexed: false, internalType: "address", name: "newVault", type: "address" },
-    ],
-    name: "SetVault",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [{ indexed: false, internalType: "address", name: "account", type: "address" }],
-    name: "Unpaused",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: true, internalType: "address", name: "to", type: "address" },
-      { indexed: false, internalType: "uint256", name: "amount", type: "uint256" },
-    ],
-    name: "Withdraw",
-    type: "event",
-  },
-  {
-    inputs: [],
-    name: "DEFAULT_ADMIN_ROLE",
-    outputs: [{ internalType: "bytes32", name: "", type: "bytes32" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "VERSION",
-    outputs: [{ internalType: "string", name: "", type: "string" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "uint256[]", name: "fids", type: "uint256[]" },
-      { internalType: "uint256", name: "units", type: "uint256" },
-    ],
-    name: "batchCredit",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "uint256[]", name: "fids", type: "uint256[]" },
-      { internalType: "uint256[]", name: "units", type: "uint256[]" },
-    ],
-    name: "batchRent",
-    outputs: [],
-    stateMutability: "payable",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "uint256", name: "start", type: "uint256" },
-      { internalType: "uint256", name: "end", type: "uint256" },
-      { internalType: "uint256", name: "units", type: "uint256" },
-    ],
-    name: "continuousCredit",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "uint256", name: "fid", type: "uint256" },
-      { internalType: "uint256", name: "units", type: "uint256" },
-    ],
-    name: "credit",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "deprecationTimestamp",
-    outputs: [{ internalType: "uint256", name: "", type: "uint256" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "ethUsdPrice",
-    outputs: [{ internalType: "uint256", name: "", type: "uint256" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "fixedEthUsdPrice",
-    outputs: [{ internalType: "uint256", name: "", type: "uint256" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [{ internalType: "bytes32", name: "role", type: "bytes32" }],
-    name: "getRoleAdmin",
-    outputs: [{ internalType: "bytes32", name: "", type: "bytes32" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "bytes32", name: "role", type: "bytes32" },
-      { internalType: "uint256", name: "index", type: "uint256" },
-    ],
-    name: "getRoleMember",
-    outputs: [{ internalType: "address", name: "", type: "address" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [{ internalType: "bytes32", name: "role", type: "bytes32" }],
-    name: "getRoleMemberCount",
-    outputs: [{ internalType: "uint256", name: "", type: "uint256" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "bytes32", name: "role", type: "bytes32" },
-      { internalType: "address", name: "account", type: "address" },
-    ],
-    name: "grantRole",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "bytes32", name: "role", type: "bytes32" },
-      { internalType: "address", name: "account", type: "address" },
-    ],
-    name: "hasRole",
-    outputs: [{ internalType: "bool", name: "", type: "bool" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "lastPriceFeedUpdateBlock",
-    outputs: [{ internalType: "uint256", name: "", type: "uint256" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "lastPriceFeedUpdateTime",
-    outputs: [{ internalType: "uint256", name: "", type: "uint256" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "maxUnits",
-    outputs: [{ internalType: "uint256", name: "", type: "uint256" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  { inputs: [], name: "pause", outputs: [], stateMutability: "nonpayable", type: "function" },
-  {
-    inputs: [],
-    name: "paused",
-    outputs: [{ internalType: "bool", name: "", type: "bool" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "prevEthUsdPrice",
-    outputs: [{ internalType: "uint256", name: "", type: "uint256" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [{ internalType: "uint256", name: "units", type: "uint256" }],
-    name: "price",
-    outputs: [{ internalType: "uint256", name: "", type: "uint256" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "priceFeed",
-    outputs: [{ internalType: "contract AggregatorV3Interface", name: "", type: "address" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "priceFeedCacheDuration",
-    outputs: [{ internalType: "uint256", name: "", type: "uint256" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "priceFeedMaxAge",
-    outputs: [{ internalType: "uint256", name: "", type: "uint256" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "priceFeedMaxAnswer",
-    outputs: [{ internalType: "uint256", name: "", type: "uint256" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "priceFeedMinAnswer",
-    outputs: [{ internalType: "uint256", name: "", type: "uint256" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  { inputs: [], name: "refreshPrice", outputs: [], stateMutability: "nonpayable", type: "function" },
-  {
-    inputs: [
-      { internalType: "bytes32", name: "role", type: "bytes32" },
-      { internalType: "address", name: "account", type: "address" },
-    ],
-    name: "renounceRole",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "uint256", name: "fid", type: "uint256" },
-      { internalType: "uint256", name: "units", type: "uint256" },
-    ],
-    name: "rent",
-    outputs: [{ internalType: "uint256", name: "overpayment", type: "uint256" }],
-    stateMutability: "payable",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "rentedUnits",
-    outputs: [{ internalType: "uint256", name: "", type: "uint256" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      { internalType: "bytes32", name: "role", type: "bytes32" },
-      { internalType: "address", name: "account", type: "address" },
-    ],
-    name: "revokeRole",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [{ internalType: "uint256", name: "duration", type: "uint256" }],
-    name: "setCacheDuration",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [{ internalType: "uint256", name: "timestamp", type: "uint256" }],
-    name: "setDeprecationTimestamp",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [{ internalType: "uint256", name: "fixedPrice", type: "uint256" }],
-    name: "setFixedEthUsdPrice",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [{ internalType: "uint256", name: "period", type: "uint256" }],
-    name: "setGracePeriod",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [{ internalType: "uint256", name: "age", type: "uint256" }],
-    name: "setMaxAge",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [{ internalType: "uint256", name: "maxPrice", type: "uint256" }],
-    name: "setMaxAnswer",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [{ internalType: "uint256", name: "max", type: "uint256" }],
-    name: "setMaxUnits",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [{ internalType: "uint256", name: "minPrice", type: "uint256" }],
-    name: "setMinAnswer",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [{ internalType: "uint256", name: "usdPrice", type: "uint256" }],
-    name: "setPrice",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [{ internalType: "contract AggregatorV3Interface", name: "feed", type: "address" }],
-    name: "setPriceFeed",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [{ internalType: "contract AggregatorV3Interface", name: "feed", type: "address" }],
-    name: "setUptimeFeed",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [{ internalType: "address", name: "vaultAddr", type: "address" }],
-    name: "setVault",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [{ internalType: "bytes4", name: "interfaceId", type: "bytes4" }],
-    name: "supportsInterface",
-    outputs: [{ internalType: "bool", name: "", type: "bool" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "unitPrice",
-    outputs: [{ internalType: "uint256", name: "", type: "uint256" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  { inputs: [], name: "unpause", outputs: [], stateMutability: "nonpayable", type: "function" },
-  {
-    inputs: [],
-    name: "uptimeFeed",
-    outputs: [{ internalType: "contract AggregatorV3Interface", name: "", type: "address" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "uptimeFeedGracePeriod",
-    outputs: [{ internalType: "uint256", name: "", type: "uint256" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "usdUnitPrice",
-    outputs: [{ internalType: "uint256", name: "", type: "uint256" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "vault",
-    outputs: [{ internalType: "address", name: "", type: "address" }],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [{ internalType: "uint256", name: "amount", type: "uint256" }],
-    name: "withdraw",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-];
