+++ packages/hub-nodejs/examples/write-data/package.json
@@ -9,10 +9,7 @@
     "typecheck": "yarn tsc --noEmit"
   },
   "dependencies": {
-    "@farcaster/hub-nodejs": "^0.10.13",
-    "axios": "^1.6.0",
-    "viem": "^1.15.4",
-    "wagmi": "^1.4.3"
+    "@farcaster/hub-nodejs": "^0.10.13"
   },
   "devDependencies": {
     "tsx": "^3.12.5",
